Move CommonCallableSupplier.
